Also fixup workspaces after setting the base branch (#7285)
---

Thanks a lot for the reproduction, I could follow it and reproduce the issue! Here is a ready-made workspace that can just be opened, ~~leading to the reproduction as soon as a commit is reordered~~, actually due to the auto-fixing behaviour implemented on application startup, one will have to follow these steps after cloning:

- `git checkout -b branch-1 origin/branch-1`
- `git checkout -b branch-2 origin/branch-2`

Alternatively, [gitbutler-stack-error.zip](https://github.com/user-attachments/files/23574086/gitbutler-stack-error.zip) is right at the state, ready to be opened for the first time.

As the auto-fixing code that matches the workspace meta-data with the actual workspace didn't run yet, the issue will reproduce.
After reopening, it won't anymore.

And I think the fix for this is to apply the auto-fixes not only when setting a project active, but also after adding it.
