+++ crates/but-api/src/legacy/mod.rs
@@ -1,0 +1,2 @@
+/// Utilities to fixup legacy code and data structures
+pub mod fixup;
